fix: use BigInteger/Integer for ID columns to fix PostgreSQL FK type mismatch
PostgreSQL rejects FK constraints between Numeric(11) columns due to
type incompatibility. Switch pure-integer columns (acct_id, cust_id,
merchant_id, cat_cd) from Numeric to BigInteger/Integer which map
correctly to PostgreSQL BIGINT/INTEGER types.

+++ modernized/app/models/account.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Numeric, String
+from sqlalchemy import BigInteger, Column, Numeric, String
 
 from app.database import Base
 
@@ -6,7 +6,7 @@ from app.database import Base
 class Account(Base):
     __tablename__ = "accounts"
 
-    acct_id = Column(Numeric(11), primary_key=True)
+    acct_id = Column(BigInteger, primary_key=True)
     active_status = Column(String(1))
     curr_bal = Column(Numeric(12, 2), default=0)
     credit_limit = Column(Numeric(12, 2), default=0)
