+++ modernized/app/models/daily_transaction.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Integer, Numeric, String
+from sqlalchemy import BigInteger, Boolean, Column, Integer, Numeric, String
 
 from app.database import Base
 
@@ -9,11 +9,11 @@ class DailyTransaction(Base):
     id = Column(Integer, primary_key=True, autoincrement=True)
     tran_id = Column(String(16), nullable=False)
     type_cd = Column(String(2))
-    cat_cd = Column(Numeric(4))
+    cat_cd = Column(Integer)
     source = Column(String(10))
     description = Column(String(100))
     amount = Column(Numeric(11, 2))
-    merchant_id = Column(Numeric(9))
+    merchant_id = Column(BigInteger)
     merchant_name = Column(String(50))
     merchant_city = Column(String(50))
     merchant_zip = Column(String(10))
