+++ modernized/app/models/transaction.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Numeric, String
+from sqlalchemy import BigInteger, Column, Integer, Numeric, String
 
 from app.database import Base
 
@@ -8,11 +8,11 @@ class Transaction(Base):
 
     tran_id = Column(String(16), primary_key=True)
     type_cd = Column(String(2))
-    cat_cd = Column(Numeric(4))
+    cat_cd = Column(Integer)
     source = Column(String(10))
     description = Column(String(100))
     amount = Column(Numeric(11, 2))
-    merchant_id = Column(Numeric(9))
+    merchant_id = Column(BigInteger)
     merchant_name = Column(String(50))
     merchant_city = Column(String(50))
     merchant_zip = Column(String(10))
